src/main/java/pro/sky/question_generator/repository/MathQuestionRepository.java
package pro.sky.question_generator.repository;

import pro.sky.question_generator.model.Question;

import java.util.Collection;
import java.util.HashSet;
import java.util.Set;

public class MathQuestionRepository implements QuestionRepository {

    private final Set<Question> questions;

    public MathQuestionRepository() {
        this.questions = new HashSet<>();
    }

    @Override
    public Question add(Question question) {
        questions.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return Set.copyOf(questions);
    }
}
